refactor: remove redundant styles footer
