--- src/app/_components/Button/Button.interface.ts
@@ -1,8 +1,0 @@
-interface ButtonProps extends React.ButtonHTMLAttributes<HTMLButtonElement> {
-  main?: boolean;
-  secondary?: boolean;
-  catalog?: boolean;
-  catalogBurgerMenu?: boolean;
-}
-
-export default ButtonProps;
