--- src/app/_components/Button/button.module.scss
@@ -1,65 +1,0 @@
-@import "@/_assets/styles/variables";
-
-.root {
-  display: flex;
-  width: 100%;
-  height: 100%;
-  align-items: center;
-  justify-content: center;
-  background-color: transparent;
-  cursor: pointer;
-  user-select: none;
-  &.main {
-    padding: 20px 10px;
-    background-color: $mainColor;
-    border: none;
-    border-radius: $borderRadiusSmall;
-    color: $light;
-    font-size: 16px;
-    font-weight: 700;
-    line-height: 1.9rem;
-    transition: background-color $transitionSet;
-    &:hover {
-      background-color: $mainColorHover;
-    }
-    &:active {
-      background-color: $mainColorActive;
-    }
-  }
-  &.secondary {
-    padding: 1.4rem 1.8rem;
-    border: 1.5px solid $mainColor;
-    border-radius: $borderRadiusSmall;
-    font-size: 1.4rem;
-    font-weight: 700;
-    line-height: 1.7rem;
-    color: $mainColor;
-    white-space: nowrap;
-  }
-  &.catalog {
-    padding: 10px 19px 10px 11px;
-    justify-content: start;
-    min-height: 46px;
-    color: $light;
-    background-color: transparent;
-    border: none;
-    outline: none;
-    font-weight: 700;
-    line-height: 2rem;
-    font-size: 16px;
-    @media(max-width: $mediaTablet) {
-      font-size: 1.2rem;
-    }
-    @media(max-width: $mediaMobile) {
-      padding: 0;
-      justify-content: center;
-    }
-  }
-  &.catalogBurgerMenu {
-    padding: 15px 20px;
-    background-color: $mainColor;
-    color: $light;
-    border: none;
-    justify-content: start;
-  }
-}
