+++ src/app/_components/Footer/footer.module.scss
@@ -6,25 +6,6 @@
   }
 }
 
-.seoInfoBtn {
-  display: flex;
-  padding: 10px 31px 10px 13px;
-  width: 100%;
-  align-items: center;
-  border: 1px solid #E9E9E9;
-  border-bottom: 2px solid #E9E9E9;
-  outline: none;
-  background-color: $light;
-  color: $mainColor;
-  line-height: 1.9rem;
-  font-weight: 700;
-  cursor: pointer;
-  border-radius: $borderRadiusMedium;
-  &.activeBtn {
-    border-bottom: 2px solid $mainColor;
-  }
-}
-
 .seoRequests {
   bottom: calc(100% + 20px);
   &.showSeo {
@@ -38,23 +19,3 @@
     min-height: 300px;
   }
 }
-
-.seoList {
-  position: absolute;
-  display: grid;
-  width: 100%;
-  height: 100%;
-  align-content: start;
-  grid-template-columns: repeat(2, 1fr);
-  grid-gap: 10px;
-  justify-items: center;
-  transform: translateX(-10000px);
-  @media(max-width: 650px) {
-    display: flex;
-    flex-direction: column;
-  }
-  &.activeList {
-    transform: translateX(0);
-  }
-}
-
