--- src/app/_components/Tabs/Tabs.interface.ts
@@ -1,12 +1,0 @@
-import { Dispatch, SetStateAction } from "react";
-
-interface TabsProps {
-  tabs: { title: string; svgIcon: string }[];
-  showIcon?: boolean;
-  activeTabIndex: number;
-  setActiveTabIndex: Dispatch<SetStateAction<number>>;
-  onTabClick?: () => void;
-  main?: boolean;
-}
-
-export default TabsProps;
